refactor: Reorganize admin hooks into resource and manage subdirectories and split authentication into login and signup hooks.
