+++ src/hooks/admin/resource/useAdminInterests.ts
